% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{close_tagging}
\alias{close_tagging}
\title{Close Tagging Session by Deleting the Tagging File}
\usage{
close_tagging(folder, tagging_file_name = "indexr_tagging.txt")
}
\arguments{
\item{folder}{A character string specifying the path to the directory containing the tagging file.}

\item{tagging_file_name}{A character string for a txt file the tagging information was saved under.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Deletes the tagging file from the specified folder, ending the tagging session.
}
\examples{
\dontrun{
close_tagging("/your/directory/path")
}
}
