% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_legacy.R
\name{update_from_legacy}
\alias{update_from_legacy}
\title{Update individual parameter files (legacy) to a single indexr.yaml parameters
file.}
\usage{
update_from_legacy(folder)
}
\arguments{
\item{folder}{the rds folder}
}
\description{
Update individual parameter files (legacy) to a single indexr.yaml parameters
file.
}
\examples{
\dontrun{
update_from_legacy(path_to_rds)
}
}
