% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hash_existence.R
\name{check_hash_existence}
\alias{check_hash_existence}
\title{Check for the Existence of Parameter Combinations in RDS Files}
\usage{
check_hash_existence(
  folder,
  parameters_list,
  halt = FALSE,
  hash_includes_timestamp = FALSE,
  ignore_na = TRUE,
  alphabetical_order = TRUE,
  algo = "xxhash64",
  ignore_script_name = FALSE
)
}
\arguments{
\item{folder}{A string specifying the directory containing the RDS files.}

\item{parameters_list}{A list of parameters for which combinations are checked.
Each element of the list should be a vector of values for one parameter.}

\item{halt}{Logical; if TRUE, the function stops execution and prints the combinations found when a specified condition ('missing' or 'existing') is met.}

\item{hash_includes_timestamp}{Logical; if TRUE, timestamps are included in the hash generation process.}

\item{ignore_na}{Logical; if TRUE, NA values are ignored during hash generation.}

\item{alphabetical_order}{Logical; if TRUE, parameters are sorted alphabetically before hash generation.}

\item{check_for}{Character vector to specify whether to check for 'missing' or 'existing' combinations.
Default options are 'missing' and 'existing'.}
}
\value{
A data frame with each row representing a parameter combination and an additional column indicating whether the combination is 'missing' or 'existing'.
}
\description{
This function checks for the existence or absence of specified parameter combinations
in RDS files within a given folder. It generates hashes for each parameter combination
and compares them against a hash table created from the RDS files.
}
\examples{
\dontrun{
folder_path <- "path/to/your/rds/files"
parameters_list <- list(param1 = c("value1", "value2"), param2 = c(1, 2))
missing_combinations <- check_parameters_existence(folder_path, parameters_list, "missing")
}
}
