% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{cleanup}
\alias{cleanup}
\title{Clean Up Untagged .rds Files in a Directory}
\usage{
cleanup(folder, cutoff_date = NULL)
}
\arguments{
\item{folder}{A character string specifying the path to the directory to clean up.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Removes any \code{.rds} files in the specified folder that are not listed in the \code{indexer_tagging.txt} tagging file.
}
\examples{
\dontrun{
# Clean up untagged .rds files in the folder
cleanup("/your/directory/path")
}
}
