% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{cleanup_from_hash_table}
\alias{cleanup_from_hash_table}
\title{Title}
\usage{
cleanup_from_hash_table(
  hash_table,
  folder,
  mode = c("manual", "all"),
  column = NULL
)
}
\arguments{
\item{column}{}
}
\description{
Title
}
