% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hash_table.R
\name{create_hash_table}
\alias{create_hash_table}
\title{Create a Hash Table from RDS Files}
\usage{
create_hash_table(folder, save_path = NULL, filter_list = NULL)
}
\arguments{
\item{folder}{A string specifying the directory containing the RDS files.}

\item{save_path}{An optional string specifying the path to save the resulting hash table as a CSV file.
If \code{NULL}, the hash table is not saved.}

\item{filter_list}{An optional list of filters to apply to the hash table.
Each element of the list should be named according to a column in the hash table
and contain the value to filter by.}
}
\value{
A data frame where each row corresponds to an \code{args_list} from an RDS file,
with an additional column for the hash of each set of arguments.
}
\description{
Scans a directory for RDS files, extracts arguments lists (\code{args_list}) from each file,
and combines them into a single data frame. Each row in the resulting data frame represents
the arguments used for one RDS file, identified by its hash. Optionally, the function can
filter the data frame based on specified criteria and save it to a file.
}
\examples{
\dontrun{
directory_path <- "path/to/your/rds/files"
hash_table <- create_hash_table(directory_path)
# To save the hash table to a file
create_hash_table(directory_path, save_path = "path/to/save/hash_table.csv")
}
}
