% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_objects.R
\name{read_objects}
\alias{read_objects}
\title{Read Objects Based on Parameter Grid}
\usage{
read_objects(
  folder,
  params_grid,
  hash_includes_timestamp = FALSE,
  ignore_na = TRUE,
  alphabetical_order = TRUE
)
}
\arguments{
\item{folder}{A string specifying the directory where the objects are stored.}

\item{params_grid}{A data frame or matrix where each row represents a different set of parameters to be hashed and checked for corresponding files.}

\item{hash_includes_timestamp}{Logical; if TRUE, includes timestamps in the hash generation process.}

\item{ignore_na}{Logical; if TRUE, NA values are ignored during hash generation.}

\item{alphabetical_order}{Logical; if TRUE, parameters are sorted alphabetically before hash generation.}
}
\value{
The function invisibly returns \code{NULL} and is primarily used for its side effect of loading objects into the global environment.
}
\description{
Reads objects from files in a specified directory. The files are identified based on a hash generated from a grid of parameters. This function iterates over each combination of parameters in the grid, generates a hash for each, and attempts to load the corresponding object from the directory.
}
\examples{
\dontrun{
folder_path <- "path/to/your/saved/objects"
params_grid <- data.frame(param1 = c("a", "b"), param2 = c(1, 2))
read_objects(folder_path, params_grid)
}
}
