% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehash.R
\name{rehash}
\alias{rehash}
\title{Rehash RDS Files in a Directory}
\usage{
rehash(
  folder,
  hash_includes_timestamp = FALSE,
  ignore_na = TRUE,
  alphabetical_order = TRUE,
  algo = "xxhash64"
)
}
\arguments{
\item{folder}{A string specifying the directory containing the RDS files to be rehashed.}

\item{hash_includes_timestamp}{Logical; if TRUE, includes timestamps in the hash generation.}

\item{ignore_na}{Logical; if TRUE, NA values are ignored during hash generation.}

\item{alphabetical_order}{Logical; if TRUE, parameters are sorted alphabetically before hash generation.}
}
\value{
The function does not return a value but renames the RDS files in the specified directory based on new hashes.
}
\description{
This function processes all RDS files in a specified directory, generating new hashes
for each file's \code{args_list} and renaming the files accordingly. It's useful when changing
the hash generation algorithm or parameters.
}
\examples{
\dontrun{
directory_path <- "path/to/your/rds/files"
rehash(directory_path)
}
}
