% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save_objects}
\alias{save_objects}
\title{Save R Objects with Parameter Hashing}
\usage{
save_objects(
  folder,
  args_list,
  ...,
  hash_includes_timestamp = FALSE,
  ignore_na = TRUE,
  alphabetical_order = TRUE,
  overwrite = FALSE,
  include_timestamp = TRUE
)
}
\arguments{
\item{folder}{A string specifying the directory where the R objects will be saved.}

\item{args_list}{A named list of arguments based on which the hash for filename will be generated.}

\item{...}{The data to be save.}

\item{hash_includes_timestamp}{Logical; if TRUE, the timestamp will be included in the hash calculation.}

\item{ignore_na}{Logical; if TRUE, NA values in \code{args_list} will be ignored for hash generation.}

\item{alphabetical_order}{Logical; if TRUE, the arguments in \code{args_list} will be sorted alphabetically before hash generation.}

\item{overwrite}{Logical; if TRUE, existing files with the same hash name will be overwritten.}

\item{include_timestamp}{Logical; if TRUE, a timestamp will be added to \code{args_list} before hash generation.}
}
\value{
Invisible NULL. This function is used for its side effect of saving files.
}
\description{
Saves a list of R objects to a file in a specified directory, with filenames
based on a hash of their parameters. Optionally includes timestamps in the
parameter list before hashing and provides options for overwriting existing
files, ignoring NA values, and sorting parameters alphabetically.
}
\examples{
\dontrun{
args_list <- list(param1 = "value1", param2 = 100)
save_objects("path/to/save", args_list)
}
}
