% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{start_tagging}
\alias{start_tagging}
\title{Start Tagging Files in a Directory}
\usage{
start_tagging(path)
}
\arguments{
\item{path}{A character string specifying the path to the directory where the tagging file will be created.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Initializes the tagging process by creating an \code{indexer_tagging.txt} file in the specified directory.
}
\examples{
\dontrun{
# Start tagging in the desired folder
start_tagging("/your/directory/path")
}
}
