% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hash_table.R
\name{update_hash_table}
\alias{update_hash_table}
\title{Update Hash Table Based on New Parameters}
\usage{
update_hash_table(
  table_path,
  rds_folder,
  hash_includes_timestamp = FALSE,
  ignore_na = TRUE,
  alphabetical_order = TRUE,
  algo = "xxhash64"
)
}
\arguments{
\item{table_path}{A string specifying the path to the CSV file containing the existing hash table.}

\item{rds_folder}{A string specifying the directory containing the RDS files associated with the hash table.}

\item{hash_includes_timestamp}{Logical; if TRUE, timestamps are included in the hash generation.}

\item{ignore_na}{Logical; if TRUE, NA values are ignored during hash generation.}

\item{alphabetical_order}{Logical; if TRUE, parameters are sorted alphabetically before hash generation.}
}
\value{
The function does not return a value but saves updated RDS files and deletes old files as needed.
}
\description{
This function updates an existing hash table by re-hashing each set of parameters with potentially updated values. It loads RDS files based on their existing hashes, updates the parameters, generates new hashes, and saves the files with the new hashes. The old files are deleted if their hashes differ from the new ones.
}
\examples{
\dontrun{
# Path to the existing hash table CSV
hash_table_path <- "path/to/hash_table.csv"
# Directory containing the associated RDS files
rds_directory <- "path/to/rds_files"
update_hash_table(hash_table_path, rds_directory)
}
}
